requirements.txt
altgraph==0.17.4
exceptiongroup==1.1.3
iniconfig==2.0.0
numpy==1.26.1
packaging==23.2
pluggy==1.3.0
pygame==2.5.2
pyinstaller==6.1.0
pyinstaller-hooks-contrib==2023.10
pytest==7.4.3
PyYAML==6.0.1
tomli==2.0.1
